+++ src/components/notification/alert-message.ts
@@ -24,7 +24,7 @@
 
 import { hostname, platform } from 'os'
 import { promisify } from 'util'
-import { format } from 'date-fns'
+import { format, formatDistanceToNow } from 'date-fns'
 import * as Handlebars from 'handlebars'
 import getos from 'getos'
 import osName from 'os-name'
@@ -40,7 +40,7 @@ import {
 
 const getLinuxDistro = promisify(getos)
 
-const getMonikaInstance = async (ipAddress: string) => {
+export const getMonikaInstance = async (ipAddress: string) => {
   const osHostname = hostname()
   await getPublicIp()
 
@@ -55,26 +55,32 @@ const getMonikaInstance = async (ipAddress: string) => {
     .join('/')})`
 }
 
-export async function getMessageForAlert({
-  alert,
-  url,
-  ipAddress,
-  probeState, // state of the probed target
-  response,
-}: {
+type MessageAlertProps = {
+  probeID: string
   alert: ProbeAlert
   url: string
   ipAddress: string
-  probeState: string
+  // state of the probed target
+  isRecovery: boolean
   response: ProbeRequestResponse
-}): Promise<NotificationMessage> {
-  const { userAgent } = getContext()
+}
+
+export async function getMessageForAlert({
+  probeID,
+  alert,
+  url,
+  ipAddress,
+  isRecovery,
+  response,
+}: MessageAlertProps): Promise<NotificationMessage> {
+  const { userAgent, incidents } = getContext()
   const [monikaInstance, osName] = await Promise.all([
     getMonikaInstance(ipAddress),
     getOSName(),
   ])
+  const recoveryOrIncident = isRecovery ? 'Recovery' : 'Incident'
   const meta = {
-    type: probeState === 'UP' ? ('recovery' as const) : ('incident' as const),
+    type: isRecovery ? ('recovery' as const) : ('incident' as const),
     url,
     time: format(new Date(), 'yyyy-MM-dd HH:mm:ss XXX'),
     hostname: hostname(),
@@ -83,19 +89,19 @@ export async function getMessageForAlert({
     monikaInstance,
     version: userAgent,
   }
-  const getSubject = (probeState: string) => {
-    const recoveryOrIncident = probeState === 'UP' ? 'Recovery' : 'Incident'
-
-    return `New ${recoveryOrIncident} from Monika`
-  }
   const getExpectedMessage = (
     alert: ProbeAlert,
-    response: ProbeRequestResponse
+    response: ProbeRequestResponse,
+    isRecovery: boolean
   ) => {
     const { status } = response
     const isHTTPStatusCode = status >= 100 && status <= 599
 
-    if (!alert.message) return ''
+    if (!alert.message) {
+      if (isRecovery)
+        return `The request is back to normal and pass the query: ${alert.query}`
+      return `The request failed because the response does not pass the query: ${alert.query}. The actual response status is ${response.status} and the response time is ${response.responseTime}.`
+    }
 
     if (!isHTTPStatusCode) {
       switch (status) {
@@ -121,8 +127,16 @@ export async function getMessageForAlert({
       },
     })
   }
-  const expectedMessage = getExpectedMessage(alert, response)
-  const bodyString = `Message: ${expectedMessage}
+  const lastIncident = incidents.find(
+    (incident) =>
+      incident.probeID === probeID && incident.probeRequestURL === url
+  )
+  const recoveryMessage = getRecoveryMessage(
+    isRecovery,
+    lastIncident?.createdAt
+  )
+  const expectedMessage = getExpectedMessage(alert, response, isRecovery)
+  const bodyString = `Message: ${recoveryMessage}${expectedMessage}
 
 URL: ${meta.url}
 
@@ -137,13 +151,29 @@ Version: ${userAgent}`
   const summary = `${expectedMessage}`
 
   const message = {
-    subject: getSubject(probeState),
+    subject: `New ${recoveryOrIncident} from Monika`,
     body: bodyString,
     summary,
     meta,
   }
 
   return message
+}
+
+function getRecoveryMessage(isRecovery: boolean, incidentDateTime?: Date) {
+  if (!isRecovery || !incidentDateTime) {
+    return ''
+  }
+
+  const incidentDuration = formatDistanceToNow(incidentDateTime, {
+    includeSeconds: true,
+  })
+  const humanReadableIncidentDateTime = format(
+    incidentDateTime,
+    'yyyy-MM-dd HH:mm:ss XXX'
+  )
+
+  return `Target is back to normal after ${incidentDuration}. The incident happened at ${humanReadableIncidentDateTime}. `
 }
 
 export const getMessageForStart = async (